% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_functions.R
\name{concordance}
\alias{concordance}
\title{Concordance}
\usage{
concordance(truth, pred)
}
\arguments{
\item{truth}{Numeric vector of ground truth assignments}

\item{pred}{Numeric vector of predicted assignments}
}
\value{
A numeric concordance score between 0 and 1
}
\description{
Calculates the concordance (overall accuracy) between ground truth and predictions.
}
\examples{
truth <- c(1, 2, 3, 0, 1000)
pred <- c(1, 2, 0, 0, 1000)
concordance(truth, pred)
}
