% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_functions.R
\name{f1_micro}
\alias{f1_micro}
\title{F1 Micro Score}
\usage{
f1_micro(truth, pred)
}
\arguments{
\item{truth}{Numeric vector of ground truth assignments (0 = negative, 1-N = singlet, 1000 = doublet)}

\item{pred}{Numeric vector of predicted assignments (0 = negative, 1-N = singlet, 1000 = doublet)}
}
\value{
A numeric F1 micro score between 0 and 1
}
\description{
Calculates the micro-averaged F1 score between ground truth and predictions.
}
\examples{
truth <- c(1, 2, 3, 0, 1000)
pred <- c(1, 2, 0, 0, 1000)
f1_micro(truth, pred)
}
